requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
flask
websocket-client
